+++ middleman/__init__.py
@@ -5,4 +5,4 @@ if sys.version_info < (3, 5):
     print("middleman requires Python 3.5 or above.")
     sys.exit(1)
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"