+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## unreleased
+## v0.2.0 - 2021-08-20
 
 ### Added